More cleanup of clang_build.
